+++ tests/unit/test_base_cache.py
@@ -36,7 +36,7 @@ def test_keys(mock_session):
     for url in [MOCKED_URL, MOCKED_URL_JSON, MOCKED_URL_REDIRECT]:
         mock_session.get(url)
 
-    all_keys = set(mock_session.cache.responses.keys()) | set(mock_session.cache.redirects.keys())
+    all_keys = set(mock_session.cache.responses.keys()) | set(mock_session.cache.aliases.keys())
     assert set(mock_session.cache.keys()) == all_keys
 
 
@@ -44,11 +44,11 @@ def test_update(mock_session):
     src_cache = BaseCache()
     for i in range(20):
         src_cache.responses[f'key_{i}'] = f'value_{i}'
-        src_cache.redirects[f'key_{i}'] = f'value_{i}'
+        src_cache.aliases[f'key_{i}'] = f'value_{i}'
 
     mock_session.cache.update(src_cache)
     assert len(mock_session.cache.responses) == 20
-    assert len(mock_session.cache.redirects) == 20
+    assert len(mock_session.cache.aliases) == 20
 
 
 def test_values(mock_session):